Enable the second abstract in native language.
It's best if the template doesn't prevent using abstracts
in both languages (english, native) when the document's
main language is English. The second abstract can be just
omitted. Support both env names, 'abstracttwo' and
'abstracten'. The latter is a bit illogical when the
first abstract is in English. Note that the terms 'first'
and 'second' do not dictate the rendering order.